snake game updated
